feat(stryker-elements): add atomic design
