--- packages/stryker-elements/src/exports/common/progress-bar.ts
@@ -1,10 +1,0 @@
-import { ProgressBar } from '../../components/common/progress-bar';
-import { defineElement } from '../../define-element';
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'progress-bar': ProgressBar;
-  }
-}
-
-defineElement('progress-bar', ProgressBar);
