--- packages/stryker-elements/src/exports/enable-button.ts
@@ -1,10 +1,0 @@
-import { defineElement } from '../define-element';
-import { EnableButton } from '../components/enable-button';
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'enable-button': EnableButton;
-  }
-}
-
-defineElement('enable-button', EnableButton);
