--- packages/stryker-elements/src/exports/hero-component.ts
@@ -1,10 +1,0 @@
-import { HeroComponent } from '../components/hero-component';
-import { defineElement } from '../define-element';
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'hero-component': HeroComponent;
-  }
-}
-
-defineElement('hero-component', HeroComponent);
