--- packages/stryker-elements/src/exports/labeled-container-component.ts
@@ -1,10 +1,0 @@
-import { LabeledContainerComponent } from '../components/labeled-container-component';
-import { defineElement } from '../define-element';
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'labeled-container-component': LabeledContainerComponent;
-  }
-}
-
-defineElement('labeled-container-component', LabeledContainerComponent);
