--- packages/stryker-elements/src/exports/layout.ts
@@ -1,10 +1,0 @@
-import { Layout } from '../components/layout';
-import { defineElement } from '../define-element';
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'layout-component': Layout;
-  }
-}
-
-defineElement('layout-component', Layout);
