+++ packages/stryker-elements/src/exports/lib/atoms/button.ts
@@ -1,0 +1,10 @@
+import { defineElement } from '../../../define-element';
+import { Button } from '../../../lib/atoms/button';
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'sme-button': Button;
+  }
+}
+
+defineElement('sme-button', Button);
