+++ packages/stryker-elements/src/exports/lib/atoms/image-button.ts
@@ -1,0 +1,10 @@
+import { defineElement } from '../../../define-element';
+import { ImageButton } from '../../../lib/atoms/image-button';
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'sme-image-button': ImageButton;
+  }
+}
+
+defineElement('sme-image-button', ImageButton);
