+++ packages/stryker-elements/src/exports/lib/atoms/link.ts
@@ -1,0 +1,10 @@
+import { defineElement } from '../../../define-element';
+import { Link } from '../../../lib/atoms/link';
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'sme-link': Link;
+  }
+}
+
+defineElement('sme-link', Link);
