+++ packages/stryker-elements/src/exports/lib/atoms/progress-bar.ts
@@ -1,0 +1,10 @@
+import { ProgressBar } from '../../../lib/atoms/progress-bar';
+import { defineElement } from '../../../define-element';
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'progress-bar': ProgressBar;
+  }
+}
+
+defineElement('progress-bar', ProgressBar);
