+++ packages/stryker-elements/src/exports/lib/molecules/profile-button.ts
@@ -1,0 +1,10 @@
+import { ProfileButton } from '../../../lib/molecules/profile-button';
+import { defineElement } from '../../../define-element';
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'sme-profile-button': ProfileButton;
+  }
+}
+
+defineElement('sme-profile-button', ProfileButton);
