+++ packages/stryker-elements/src/exports/lib/molecules/repository.ts
@@ -1,0 +1,10 @@
+import { Repository } from '../../../lib/molecules/repository';
+import { defineElement } from '../../../define-element';
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'sme-repository': Repository;
+  }
+}
+
+defineElement('sme-repository', Repository);
