--- packages/stryker-elements/src/exports/status-box-component.ts
@@ -1,10 +1,0 @@
-import { StatusBoxComponent } from '../components/status-box-component';
-import { defineElement } from '../define-element';
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'status-box-element': StatusBoxComponent;
-  }
-}
-
-defineElement('status-box-component', StatusBoxComponent);
