+++ packages/stryker-elements/src/main.ts
@@ -1,35 +1,44 @@
-export { HeroComponent } from './components/hero-component';
-import './exports/hero-component';
+// Atoms
 
-export { ProfileButton } from './components/profile-button';
-import './exports/profile-button';
+export { Button } from './lib/atoms/button';
+import './exports/lib/atoms/button';
 
-export { EnableButton } from './components/enable-button';
-import './exports/enable-button';
+export { ImageButton } from './lib/atoms/image-button';
+import './exports/lib/atoms/image-button';
 
-export { Button } from './components/button';
-import './exports/button';
+export { LabeledContainer } from './lib/atoms/labeled-container';
+import './exports/lib/atoms/labeled-container';
 
-export { SignInButton } from './components/sign-in-button';
-import './exports/sign-in-button';
+export { Link } from './lib/atoms/link';
+import './exports/lib/atoms/link';
 
-export { GettingStartedStepOverview } from './components/getting-started/getting-started-step-overview';
-import './exports/getting-started/getting-started-step-overview';
+export { ProgressBar } from './lib/atoms/progress-bar';
+import './exports/lib/atoms/progress-bar';
 
-export { ProgressBar } from './components/common/progress-bar';
-import './exports/common/progress-bar';
+export { SpatiousLayout } from './lib/atoms/spatious-layout';
+import './exports/lib/atoms/spatious-layout';
 
-export { ResultItem } from './components/common/result-item';
-import './exports/common/result-item';
+// Molecules
 
-export { TopBar } from './components/top-bar';
-import './exports/top-bar';
+export { GettingStartedStep } from './lib/molecules/getting-started-step';
+import './exports/lib/molecules/getting-started-step';
 
-export { StatusBoxComponent } from './components/status-box-component';
-import './exports/status-box-component';
+export { ProfileButton } from './lib/molecules/profile-button';
+import './exports/lib/molecules/profile-button';
 
-export { Layout } from './components/layout';
-import './exports/layout';
+export { Repository } from './lib/molecules/repository';
+import './exports/lib/molecules/repository';
 
-export { LabeledContainerComponent } from './components/labeled-container-component';
-import './exports/labeled-container-component';
+// Organisms
+
+export { GettingStartedOverview } from './lib/organisms/getting-started-overview';
+import './exports/lib/organisms/getting-started-overview';
+
+export { Hero } from './lib/organisms/hero';
+import './exports/lib/organisms/hero';
+
+export { StatusOverview } from './lib/organisms/status-overview';
+import './exports/lib/organisms/status-overview';
+
+export { TopBar } from './lib/organisms/top-bar';
+import './exports/lib/organisms/top-bar';
