+++ packages/stryker-elements/src/tailwind-styles/global.css
