+++ packages/website-frontend/src/app/app.component.html
@@ -1,6 +1,6 @@
-<top-bar logo="https://stryker-mutator.io/images/stryker.svg">
-  <stryker-user slot="user"></stryker-user>
-</top-bar>
+<sme-top-bar logoUrl="https://stryker-mutator.io/images/stryker.svg">
+  <stryker-user slot="right-side"></stryker-user>
+</sme-top-bar>
 <div>
   <router-outlet></router-outlet>
 </div>
