+++ packages/website-frontend/src/app/welcome/welcome.component.html
@@ -1,4 +1,4 @@
-<hero-component></hero-component>
-<layout-component>
-  <getting-started-step-overview></getting-started-step-overview>
-</layout-component>
+<sme-hero></sme-hero>
+<sme-spatious-layout>
+  <sme-getting-started-overview id="getting-started"></sme-getting-started-overview>
+</sme-spatious-layout>
